+++ wiki/Changelog.md
@@ -1,3 +1,23 @@
+### 2.7.3 (23 Dec 2016) ###
+
+Character display:
+  * Tweaked surrogate handling to facilitate non-BMP display (#616).
+
+Localization:
+  * Persisting adaptation of font sample text.
+  * Disambiguated translations in different context for "Paste", "Font", "Colour".
+  * Sorted out en_GB vs. en_US.
+  * Fixed localization of Options dialog window title.
+
+### 2.7.2 (10 Dec 2016) ###
+
+Localization:
+  * Completion of localization support, covering font chooser and colour chooser (#537).
+  * Hint in Theme field as feedback after dragging downloaded colour scheme.
+
+Other:
+  * Fix config dialog crash when trying to load resources from network drive (#610).
+
 ### 2.7.1 (4 Dec 2016) ###
 
 Character display: